+++ aiosmtpd/docs/NEWS.rst
@@ -2,9 +2,13 @@
  NEWS for aiosmtpd
 ===================
 
-2.0 (201X-XX-XX)
+1.1 (201X-XX-XX)
 ================
 * Drop support for Python 3.4.
+* When aiosmtpd handles a ``STARTTLS`` it must arrange for the original
+  transport to be closed when the wrapped transport is closed.  This fixes a
+  hidden exception which occurs when an EOF is received on the original
+  tranport after the connection is lost.  (Closes #83)
 
 1.0 (2017-05-15)
 ================